UI Main Menu by Jakub Szala
I've added my version in Main Menu in a seperate folder, as there was another menu already. It was a different type of node so I've decided not to change original one to avoid dev's code design idea or something.

MainMenu contains a placeholder design by me, I did it just for fun and to test Godot features, I do not oppose changing it in any way. I added a theme for Menu UI and some styleboxes to it and a free font I downloaded from the web. The placeholder background image is bing AI generated and it should be replaced cuz of ethical reasons and it's horryfing visages, but I wanted to add it to make this look cooler and more complete.

MainMenu scene contains two ColoredRect nodes, one of which has title texture placeholder if needed for graphics team if they decide to create seperate art for a title. They can just delete the ColoredRect and leave TextureRect if they decide to do so.